test(registry): the user-registry suite has one owner, and its alias pin is not inert
tests/test_user_registry.py and tests/registry/test_user_registry.py were two
copies of the same suite for strands_robots.registry.user_registry. 35 cells
ran twice; 4 behaviours were pinned only at the root path and 1 only under
tests/registry/, so a behaviour change had to be made in two places and each
copy kept its own version.

The drift produced contradictory contracts. The root copy pinned the shipped
fail-closed rule (register_robot refuses an alias that would collide at read
time). The copy in the canonical directory pinned a fail-soft path that no
longer exists: it patches registry.robots.list_robots to raise and asserts
registration survives, but register_robot() does not call list_robots -
measured on main, the mock's call_count is 0 - and under that same premise a
colliding alias still raises ValueError. The cell passed vacuously while
documenting the opposite of the behaviour.

Keep one module, under tests/registry/ next to the module it covers and inside
that directory's isolation conftest. Carry over the alias-collision class and
the missing-asset-dir pin, drop the inert cell. 76 collected cells become 40;
statement coverage of user_registry.py is unchanged at 93 statements / 10
missed / 89%.
