+++ tests/registry/test_user_registry.py
@@ -106,46 +106,6 @@ class TestRegisterRobot:
         assert entry["hardware"] == hw
 
 
-class TestRegisterRobotAliasCollisionFailSoft:
-    """Alias-collision detection must degrade gracefully, never block a
-    registration.
-
-    ``register_robot(aliases=...)`` enumerates the package registry to warn
-    when a user alias shadows a canonical name or another alias. That
-    enumeration is best-effort diagnostics, not a precondition: if listing
-    the package registry raises (corrupt/partial install, import-time error
-    in a downstream ``robots.py`` consumer), registration must still succeed
-    with the aliases intact rather than propagating the failure to the
-    caller. This pins the ``except Exception`` fallback that resets the
-    package canonical/alias sets to empty.
-    """
-
-    def test_register_with_aliases_survives_package_enumeration_failure(self, tmp_path):
-        """When list_robots() raises during alias-collision detection, the
-        robot is still registered and its aliases resolve."""
-        robot_dir = _make_robot(tmp_path / "assets")
-
-        # The collision-detection block does a function-local
-        # ``from .robots import list_robots`` then calls it; patch the name
-        # on the robots module so the call raises at enumeration time.
-        with mock.patch(
-            "strands_robots.registry.robots.list_robots",
-            side_effect=RuntimeError("package registry enumeration failed"),
-        ):
-            entry = register_robot(
-                name="failsoft_bot",
-                model_xml="bot.xml",
-                asset_dir=str(robot_dir),
-                aliases=["failsoft_alias"],
-            )
-
-        # Registration succeeded despite the enumeration failure ...
-        assert entry["aliases"] == ["failsoft_alias"]
-        # ... and the robot + its alias are actually persisted/resolvable.
-        assert resolve_name("failsoft_alias") == "failsoft_bot"
-        assert get_user_robots().get("failsoft_bot") is not None
-
-
 class TestRegisterRobotNameNormalization:
     """Names are lower-cased, stripped, and hyphens become underscores."""
 
@@ -193,6 +153,56 @@ class TestRegisterRobotValidation:
         empty_dir.mkdir(parents=True)
         with pytest.raises(FileNotFoundError, match="Model XML not found"):
             register_robot(name="empty", model_xml="nope.xml", asset_dir=str(empty_dir))
+
+    def test_missing_asset_dir_raises_file_not_found(self, tmp_path):
+        """A non-existent asset_dir fails closed at registration with an actionable message.
+
+        Registration must not silently accept a dir that does not exist yet and
+        defer the failure to add_robot() time.
+        """
+        ghost_dir = tmp_path / "assets" / "ghost"  # never created
+        with pytest.raises(FileNotFoundError, match="Asset directory does not exist") as exc:
+            register_robot(name="ghost", model_xml="g.xml", asset_dir=str(ghost_dir))
+        # Message names the missing dir and the xml the user should place inside it.
+        assert str(ghost_dir) in str(exc.value)
+        assert "g.xml" in str(exc.value)
+        # Nothing should have been persisted.
+        assert get_user_robots().get("ghost") is None
+
+
+class TestRegisterRobotAliasCollision:
+    """register_robot fails closed when an alias would collide at read time.
+
+    The loader RAISES on an alias that shadows a canonical name or another
+    robot's alias on every subsequent read, so register_robot must reject such
+    an alias at write time rather than persist it. Otherwise a "successful"
+    registration bricks every get_robot/resolve_name call process-wide until
+    user_robots.json is hand-edited.
+    """
+
+    def test_alias_shadowing_package_canonical_name_raises(self, tmp_path):
+        robot_dir = _make_robot(tmp_path / "assets", name="botA", xml_name="a.xml")
+        with pytest.raises(ValueError, match="so100"):
+            register_robot(name="botA", model_xml="a.xml", asset_dir=str(robot_dir), aliases=["so100"])
+        # Nothing persisted, and the registry still loads.
+        assert get_user_robots().get("bota") is None
+        assert get_robot("so100") is not None
+
+    def test_alias_already_used_by_another_user_robot_raises(self, tmp_path):
+        first_dir = _make_robot(tmp_path / "assets", name="botB", xml_name="b.xml")
+        second_dir = _make_robot(tmp_path / "assets", name="botC", xml_name="c.xml")
+        register_robot(name="botB", model_xml="b.xml", asset_dir=str(first_dir), aliases=["grabber"])
+        with pytest.raises(ValueError, match="grabber"):
+            register_robot(name="botC", model_xml="c.xml", asset_dir=str(second_dir), aliases=["grabber"])
+        # The second robot is not persisted; the first is still resolvable.
+        assert get_user_robots().get("botc") is None
+        assert get_robot("grabber") is not None
+
+    def test_unique_alias_registers_successfully(self, tmp_path):
+        robot_dir = _make_robot(tmp_path / "assets", name="botD", xml_name="d.xml")
+        register_robot(name="botD", model_xml="d.xml", asset_dir=str(robot_dir), aliases=["unique_grip_xyz"])
+        assert get_user_robots().get("botd") is not None
+        assert get_robot("unique_grip_xyz") is not None
 
 
 class TestRegisterRobotAssetDirResolution:
